ai-serving-sync/requirements.txt
#pytorch
#from https://download.pytorch.org/whl/torch/
https://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl#sha256=1f8efaebfcbb7ec3962fd8c7c3be02c6666eff53a12043006a749d647656163e
#from https://download.pytorch.org/whl/torchvision/
https://download.pytorch.org/whl/cu118/torchvision-0.15.0%2Bcu118-cp38-cp38-linux_x86_64.whl#sha256=03a67002cdae2f4cc8316358f0627d35e8aa1d61be4595e7c9fbea3ac16b7c00

uvicorn
pillow
fastapi
pycocotools
randaugment
python-dotenv
nest-asyncio